Add update_only to cache.update_instance
This can change the behaviour when then instance is not already cached.

With update_only=False (default, old behaviour), uncached instances are
cached, and follow-on invalidators are returned.  If a client is
recursively calling update_instance, this can load a lot of instances
into a cold cache.

With update_only=True, uncached instances are not cached, and an empty
invalidation list is returned.  If a client is recursively calling
update_instance, this will quickly halt instance loading for a cold
cache, but may result in an inconsistent cache.  This is suitable for
development or when eventual consistency is appropriate.

In either case, if the instance is already cached, it will be checked to
see if it needs updating.