Allow null and strings as date values
When creating or modifying Django instances, it is possible to use a
string for DateField, DateTimeField, and DurationField values.  Before
writing to the database, these are converted to Date, DateTime, etc.,
but when reading the value from the new instance, it will still be a
string. This code uses the Django utilities to parse strings into
datetime objects before serializing them to JSON for the cache.

For nullable fields, None is a possible value, and this is serialized as
null for the cache.